Move ordereddict backport to _bundled directory.
