+++ vispy/ext/ordereddict.py
@@ -4,4 +4,4 @@ from sys import version_info
 if version_info >= (2, 7):
     from collections import OrderedDict
 else:
-    from .py24_ordereddict import OrderedDict  # noqa
+    from ._bundled.ordereddict import OrderedDict  # noqa
